Project Description
This is the repository for Filippos Lymperopoulos and James Jang's final project for the class "Mechanics: The ModSim/Experimental Approach", Spring 2014, at Franklin W. Olin College of Engineering. We studied the sport of basketball and, more precisely, the nature of swishshots and banked shots. We developed a 3D model of the sport and conducted motion tracking to validate our model. We further considered rim and backboard interactions and developed some very important figures of merit, including frontspin, backspin, and no spin at all.